Create Athlete_Sort.py
You are given a spreadsheet that contains a list of N athletes and their details (such as age, height, weight and so on). You are required to sort the data based on the Kth attribute and print the final resulting table. Follow the example given below for better understanding.

Rank Age Height (in

cm)

Rark

Age Height (in cm)

1

32

190 175

5

24

176

35

son based on t

4

26

195

2

3

41

188

1

32

190

4

26

195

2

35

3 41

175

188

5

24 176

Note that K is indexed from 0 to M-1, where M is the number of attributes.

Note: If two attributes are the same for different rows, for example, if two atheletes are of the same age, print the

row that appeared first in the input.

Input Format

The first line contains N and M separated by a space.

The next N lines each contain M elements.

The last line contains K.

Constraints

1<N, M<1000 0≤K<M

Each element 1000
Print the N lines of the sorted table. Each line should contain the space separated elements. Check the sample

below for clarity.

Sample Input 0

53

10 25

718

999 123 12

659

1

Sample Output 0

710

18 25

659

999

1 23 12

>

Explanation 0

The details are sorted based on the second attribute, since K is zero-indexed.